Move Frustum class into math utils
